backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
openai>=2.26.0,<3
sentence-transformers==3.0.1
numpy==1.26.4
scikit-learn==1.5.0
scipy==1.13.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
pydantic>=2.7.1,<3
python-multipart==0.0.9
APScheduler==3.10.4
icalendar>=5.0.11,<8
recurring-ical-events>=3.6.0,<4
sqlalchemy>=2.0.0,<3
bcrypt>=4.0.0,<5
PyJWT>=2.8.0,<3
langgraph>=1.0,<2
langchain-openai>=1.0,<2
langchain-core>=1.0,<2

# 可选：chyroc/wechatsogou（搜狗微信）；需 cachelib 顶替已移除的 werkzeug.contrib.cache
cachelib>=0.13,<1
wechatsogou==4.5.4